Update page index -> home & arg incomePerYear -> incomesPerYear
